+++ NEWS.md
@@ -1,3 +1,9 @@
+# lavacreg 0.1-1
+
+* skipping two tests failing on MacOS (i.e. applications in which only one latent covariate is considered)
+* lowered numerical precision for test on Poisson regression with three manifest covariates (for Solaris)
+* added a more detailed summary-function
+
 # lavacreg 0.1-0
 
 * Added a `NEWS.md` file to track changes to the package.